+++ src/routes/layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { Header } from "components/header";
+import { Header } from "components/Header";
 import { Box } from "@mui/material";
 
 export const Layout = () => (
